+++ app/Werewolf/Commands/Vote.hs
@@ -19,19 +19,16 @@ module Werewolf.Commands.Vote (
     handle,
 ) where
 
-import Control.Lens
 import Control.Monad.Except
 import Control.Monad.Extra
 import Control.Monad.State
 import Control.Monad.Writer
 
-import Data.Maybe
-import Data.Text  (Text)
+import Data.Text (Text)
 
 import Game.Werewolf.Command
-import Game.Werewolf.Engine
+import Game.Werewolf.Engine   hiding (isVillagersTurn)
 import Game.Werewolf.Game
-import Game.Werewolf.Player   as Player
 import Game.Werewolf.Response
 
 -- | Options.
@@ -42,15 +39,17 @@ data Options = Options
 -- | Handle.
 handle :: MonadIO m => Text -> Options -> m ()
 handle callerName (Options targetName) = do
-    unlessM doesGameExist $ exitWith failure { messages = [privateMessage [callerName] "No game is running."] }
+    unlessM doesGameExist $ exitWith failure {
+        messages = [privateMessage [callerName] "No game is running."]
+        }
 
     game <- readGame
 
-    let mCaller = findByName callerName (game ^. players)
-    when (isNothing mCaller) $ exitWith failure { messages = [playerDoesNotExistMessage callerName callerName] }
+    let command = (if isVillagersTurn game
+            then lynchVoteCommand
+            else killVoteCommand
+            ) callerName targetName
 
-    let command = (if isVillager (fromJust mCaller) then lynchVoteCommand else killVoteCommand) callerName targetName
-
-    case runExcept (runWriterT $ execStateT (apply command >> advanceTurn >> checkGameOver) game) of
+    case runExcept (runWriterT $ execStateT (apply command >> checkTurn >> checkGameOver) game) of
         Left errorMessages      -> exitWith failure { messages = errorMessages }
         Right (game', messages) -> writeGame game' >> exitWith success { messages = messages }