+++ test/src/Game/Werewolf/Test/Engine.hs
@@ -9,17 +9,16 @@ Maintainer  : public@hjwylde.com
 {-# LANGUAGE OverloadedStrings #-}
 
 module Game.Werewolf.Test.Engine (
-    -- * advanceTurn
-    prop_advanceTurnSkipsSeersWhenNoSeers, prop_advanceTurnDoesNothingWhenGameOver,
-    prop_advanceSeersTurnAdvancesToWerewolves, prop_advanceSeersTurnResetsSees,
-    prop_advanceSeersTurnDoesNothingUnlessAllSeen, prop_advanceVillagersTurnAdvancesToSeers,
-    prop_advanceVillagersTurnLynchesOnePlayerWhenConsensus,
-    prop_advanceVillagersTurnLynchesNoOneWhenConflicted, prop_advanceVillagersTurnResetsVotes,
-    prop_advanceVillagersTurnDoesNothingUnlessAllVoted,
-    prop_advanceWerewolvesTurnAdvancesToVillagers,
-    prop_advanceWerewolvesTurnKillsOnePlayerWhenConsensus,
-    prop_advanceWerewolvesTurnKillsNoOneWhenConflicted, prop_advanceWerewolvesTurnResetsVotes,
-    prop_advanceWerewolvesTurnDoesNothingUnlessAllVoted,
+    -- * checkTurn
+    prop_checkTurnSkipsSeersWhenNoSeers, prop_checkTurnDoesNothingWhenGameOver,
+    prop_checkSeersTurnAdvancesToWerewolves, prop_checkSeersTurnResetsSees,
+    prop_checkSeersTurnDoesNothingUnlessAllSeen, prop_checkVillagersTurnAdvancesToSeers,
+    prop_checkVillagersTurnLynchesOnePlayerWhenConsensus,
+    prop_checkVillagersTurnLynchesNoOneWhenConflicted, prop_checkVillagersTurnResetsVotes,
+    prop_checkVillagersTurnDoesNothingUnlessAllVoted, prop_checkWerewolvesTurnAdvancesToVillagers,
+    prop_checkWerewolvesTurnKillsOnePlayerWhenConsensus,
+    prop_checkWerewolvesTurnKillsNoOneWhenConflicted, prop_checkWerewolvesTurnResetsVotes,
+    prop_checkWerewolvesTurnDoesNothingUnlessAllVoted,
 
     -- * checkGameOver
     prop_checkGameOverAdvancesTurn, prop_checkGameOverDoesNothingWhenAtLeastTwoAllegiancesAlive,
@@ -39,6 +38,7 @@ module Game.Werewolf.Test.Engine (
 
 import Control.Lens         hiding (elements)
 import Control.Monad.Except
+import Control.Monad.Writer
 
 import           Data.Either.Extra
 import           Data.List.Extra
@@ -55,124 +55,124 @@ import Game.Werewolf.Test.Arbitrary
 import Test.QuickCheck
 import Test.QuickCheck.Monadic
 
-prop_advanceTurnSkipsSeersWhenNoSeers :: Game -> Property
-prop_advanceTurnSkipsSeersWhenNoSeers game =
+prop_checkTurnSkipsSeersWhenNoSeers :: Game -> Property
+prop_checkTurnSkipsSeersWhenNoSeers game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    isWerewolvesTurn $ run_ advanceTurn game''
+    isWerewolvesTurn $ run_ checkTurn game''
     where
         game'   = (foldl killPlayer game (filterSeers $ game ^. players)) { _turn = Villagers }
         n       = length . filterAlive $ game' ^. players
 
-prop_advanceTurnDoesNothingWhenGameOver :: Game -> Property
-prop_advanceTurnDoesNothingWhenGameOver game = run_ advanceTurn game' === game'
+prop_checkTurnDoesNothingWhenGameOver :: Game -> Property
+prop_checkTurnDoesNothingWhenGameOver game = run_ checkTurn game' === game'
     where
         game' = game { _turn = NoOne }
 
-prop_advanceSeersTurnAdvancesToWerewolves :: Game -> Property
-prop_advanceSeersTurnAdvancesToWerewolves game =
+prop_checkSeersTurnAdvancesToWerewolves :: Game -> Property
+prop_checkSeersTurnAdvancesToWerewolves game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    isWerewolvesTurn $ run_ advanceTurn game''
+    isWerewolvesTurn $ run_ checkTurn game''
     where
         game'   = game { _turn = Seers }
         n       = length . filterSeers $ game' ^. players
 
-prop_advanceSeersTurnResetsSees :: Game -> Property
-prop_advanceSeersTurnResetsSees game =
+prop_checkSeersTurnResetsSees :: Game -> Property
+prop_checkSeersTurnResetsSees game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    Map.null $ run_ advanceTurn game'' ^. sees
+    Map.null $ run_ checkTurn game'' ^. sees
     where
         game'   = game { _turn = Seers }
         n       = length . filterSeers $ game' ^. players
 
-prop_advanceSeersTurnDoesNothingUnlessAllSeen :: Game -> Property
-prop_advanceSeersTurnDoesNothingUnlessAllSeen game =
+prop_checkSeersTurnDoesNothingUnlessAllSeen :: Game -> Property
+prop_checkSeersTurnDoesNothingUnlessAllSeen game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    isSeersTurn $ run_ advanceTurn game''
+    isSeersTurn $ run_ checkTurn game''
     where
         game'   = game { _turn = Seers }
         n       = length (filterSeers $ game' ^. players) - 1
 
-prop_advanceVillagersTurnAdvancesToSeers :: Game -> Property
-prop_advanceVillagersTurnAdvancesToSeers game =
+prop_checkVillagersTurnAdvancesToSeers :: Game -> Property
+prop_checkVillagersTurnAdvancesToSeers game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    not (null . filterAlive . filterSeers $ run_ advanceTurn game'' ^. players)
-    ==> isSeersTurn $ run_ advanceTurn game''
+    not (null . filterAlive . filterSeers $ run_ checkTurn game'' ^. players)
+    ==> isSeersTurn $ run_ checkTurn game''
     where
         game'   = game { _turn = Villagers }
         n       = length $ game' ^. players
 
-prop_advanceVillagersTurnLynchesOnePlayerWhenConsensus :: Game -> Property
-prop_advanceVillagersTurnLynchesOnePlayerWhenConsensus game =
+prop_checkVillagersTurnLynchesOnePlayerWhenConsensus :: Game -> Property
+prop_checkVillagersTurnLynchesOnePlayerWhenConsensus game =
     forAll (runArbitraryCommands n game') $ \game'' ->
     length (last $ groupSortOn (length . flip elemIndices (Map.elems $ game'' ^. votes)) (nub . Map.elems $ game'' ^. votes)) == 1
-    ==> length (filterDead $ run_ advanceTurn game'' ^. players) == 1
+    ==> length (filterDead $ run_ checkTurn game'' ^. players) == 1
     where
         game'   = game { _turn = Villagers }
         n       = length $ game' ^. players
 
-prop_advanceVillagersTurnLynchesNoOneWhenConflicted :: Game -> Property
-prop_advanceVillagersTurnLynchesNoOneWhenConflicted game =
+prop_checkVillagersTurnLynchesNoOneWhenConflicted :: Game -> Property
+prop_checkVillagersTurnLynchesNoOneWhenConflicted game =
     forAll (runArbitraryCommands n game') $ \game'' ->
     length (last $ groupSortOn (length . flip elemIndices (Map.elems $ game'' ^. votes)) (nub . Map.elems $ game'' ^. votes)) > 1
-    ==> length (filterDead $ run_ advanceTurn game'' ^. players) == 0
+    ==> length (filterDead $ run_ checkTurn game'' ^. players) == 0
     where
         game'   = game { _turn = Villagers }
         n       = length $ game' ^. players
 
-prop_advanceVillagersTurnResetsVotes :: Game -> Property
-prop_advanceVillagersTurnResetsVotes game =
+prop_checkVillagersTurnResetsVotes :: Game -> Property
+prop_checkVillagersTurnResetsVotes game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    Map.null $ run_ advanceTurn game'' ^. votes
+    Map.null $ run_ checkTurn game'' ^. votes
     where
         game'   = game { _turn = Villagers }
         n       = length $ game' ^. players
 
-prop_advanceVillagersTurnDoesNothingUnlessAllVoted :: Game -> Property
-prop_advanceVillagersTurnDoesNothingUnlessAllVoted game =
+prop_checkVillagersTurnDoesNothingUnlessAllVoted :: Game -> Property
+prop_checkVillagersTurnDoesNothingUnlessAllVoted game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    isVillagersTurn $ run_ advanceTurn game''
+    isVillagersTurn $ run_ checkTurn game''
     where
         game'   = game { _turn = Villagers }
         n       = length (game' ^. players) - 1
 
-prop_advanceWerewolvesTurnAdvancesToVillagers :: Game -> Property
-prop_advanceWerewolvesTurnAdvancesToVillagers game =
+prop_checkWerewolvesTurnAdvancesToVillagers :: Game -> Property
+prop_checkWerewolvesTurnAdvancesToVillagers game =
     forAll (runArbitraryCommands n game') $ \game' ->
-    isVillagersTurn $ run_ advanceTurn game'
+    isVillagersTurn $ run_ checkTurn game'
     where
         game'   = game { _turn = Werewolves }
         n       = length . filterWerewolves $ game' ^. players
 
-prop_advanceWerewolvesTurnKillsOnePlayerWhenConsensus :: Game -> Property
-prop_advanceWerewolvesTurnKillsOnePlayerWhenConsensus game =
+prop_checkWerewolvesTurnKillsOnePlayerWhenConsensus :: Game -> Property
+prop_checkWerewolvesTurnKillsOnePlayerWhenConsensus game =
     forAll (runArbitraryCommands n game') $ \game'' ->
     length (last $ groupSortOn (length . flip elemIndices (Map.elems $ game'' ^. votes)) (nub . Map.elems $ game'' ^. votes)) == 1
-    ==> length (filterDead $ run_ advanceTurn game'' ^. players) == 1
+    ==> length (filterDead $ run_ checkTurn game'' ^. players) == 1
     where
         game'   = game { _turn = Werewolves }
         n       = length $ game' ^. players
 
-prop_advanceWerewolvesTurnKillsNoOneWhenConflicted :: Game -> Property
-prop_advanceWerewolvesTurnKillsNoOneWhenConflicted game =
+prop_checkWerewolvesTurnKillsNoOneWhenConflicted :: Game -> Property
+prop_checkWerewolvesTurnKillsNoOneWhenConflicted game =
     forAll (runArbitraryCommands n game') $ \game'' ->
     length (last $ groupSortOn (length . flip elemIndices (Map.elems $ game'' ^. votes)) (nub . Map.elems $ game'' ^. votes)) > 1
-    ==> length (filterDead $ run_ advanceTurn game'' ^. players) == 0
+    ==> length (filterDead $ run_ checkTurn game'' ^. players) == 0
     where
         game'   = game { _turn = Werewolves }
         n       = length $ game' ^. players
 
-prop_advanceWerewolvesTurnResetsVotes :: Game -> Property
-prop_advanceWerewolvesTurnResetsVotes game =
+prop_checkWerewolvesTurnResetsVotes :: Game -> Property
+prop_checkWerewolvesTurnResetsVotes game =
     forAll (runArbitraryCommands n game') $ \game' ->
-    Map.null $ run_ advanceTurn game' ^. votes
+    Map.null $ run_ checkTurn game' ^. votes
     where
         game'   = game { _turn = Werewolves }
         n       = length . filterWerewolves $ game' ^. players
 
-prop_advanceWerewolvesTurnDoesNothingUnlessAllVoted :: Game -> Property
-prop_advanceWerewolvesTurnDoesNothingUnlessAllVoted game =
+prop_checkWerewolvesTurnDoesNothingUnlessAllVoted :: Game -> Property
+prop_checkWerewolvesTurnDoesNothingUnlessAllVoted game =
     forAll (runArbitraryCommands n game') $ \game'' ->
-    isWerewolvesTurn $ run_ advanceTurn game''
+    isWerewolvesTurn $ run_ checkTurn game''
     where
         game'   = game { _turn = Werewolves }
         n       = length (filterWerewolves $ game' ^. players) - 1
@@ -194,28 +194,28 @@ prop_checkGameOverDoesNothingWhenAtLeastTwoAllegiancesAlive game =
 
 prop_startGameStartsWithSeersTurn :: [Player] -> Property
 prop_startGameStartsWithSeersTurn players = and [
-    isRight . runExcept $ startGame "" players
-    ] ==> isSeersTurn (fromRight . runExcept $ startGame "" players)
+    isRight . runExcept . runWriterT $ startGame "" players
+    ] ==> isSeersTurn (fst . fromRight . runExcept . runWriterT $ startGame "" players)
 
 prop_startGameUsesGivenPlayers :: [Player] -> Property
 prop_startGameUsesGivenPlayers players' = and [
-    isRight . runExcept $ startGame "" players'
-    ] ==> (fromRight . runExcept $ startGame "" players') ^. players == players'
+    isRight . runExcept . runWriterT $ startGame "" players'
+    ] ==> (fst . fromRight . runExcept . runWriterT $ startGame "" players') ^. players == players'
 
 prop_startGameErrorsUnlessUniquePlayerNames :: [Player] -> Property
 prop_startGameErrorsUnlessUniquePlayerNames players = and [
-    isRight . runExcept $ startGame "" players
-    ] ==> forAll (elements players) $ \player -> isLeft (runExcept $ startGame "" (player:players))
+    isRight . runExcept . runWriterT $ startGame "" players
+    ] ==> forAll (elements players) $ \player -> isLeft (runExcept . runWriterT $ startGame "" (player:players))
 
 prop_startGameErrorsWhenLessThan7Players :: [Player] -> Property
 prop_startGameErrorsWhenLessThan7Players players = and [
     length players < 7
-    ] ==> isLeft (runExcept $ startGame "" players)
+    ] ==> isLeft (runExcept . runWriterT $ startGame "" players)
 
 prop_startGameErrorsWhenMoreThan24Players :: Property
 prop_startGameErrorsWhenMoreThan24Players = forAll (resize 30 $ listOf arbitrary) $ \players -> and [
     length players > 24
-    ] ==> isLeft (runExcept $ startGame "" players)
+    ] ==> isLeft (runExcept . runWriterT $ startGame "" players)
 
 prop_createPlayersUsesGivenPlayerNames :: [Text] -> Property
 prop_createPlayersUsesGivenPlayerNames playerNames = monadicIO $ createPlayers playerNames >>= return . (playerNames ==) . map _name